# documentation/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /v1/users:
    post:
      tags:
      - users-controller
      operationId: postUsers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        "200":
          description: OK
  /v1/counters:
    get:
      tags:
      - counters-controller
      operationId: getCounters
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CounterDto'
    post:
      tags:
      - counters-controller
      operationId: postCounters
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CounterDto'
        required: true
      responses:
        "200":
          description: OK
  /v1/counters/{name}:
    get:
      tags:
      - counters-controller
      operationId: getCounter
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CounterDto'
    post:
      tags:
      - counters-controller
      operationId: postCounters_1
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: operation
        in: query
        required: true
        schema:
          type: string
          enum:
          - INCREMENT
      responses:
        "200":
          description: OK
components:
  schemas:
    UserDto:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    CounterDto:
      type: object
      properties:
        name:
          type: string
        value:
          type: integer
          format: int64